requirements.txt
Django==2.1.10
djangorestframework==3.10.0
Mako==1.0.13
Markdown==3.1.1
psycopg2==2.8.3
requests==2.20.1
six==1.11.0
dj-database-url==0.5.0
gunicorn==19.9.0 
whitenoise==4.1.3
